+++ example/lib/example.dart
@@ -114,8 +114,8 @@ void main(List<String> arguments) async {
 
   final VideoFull shionSingingStream = await holodexClient.getVideoMetadata(
     'eJJuy5rY57w', // Shion's singing stream
-    timestampComments: true,
-    recommendationLanguages: [Language.all],
+    includeTimestampComments: true,
+    languages: [Language.all],
   );
   final List<Comment>? timestampComments = shionSingingStream.comments;
   final List<Video>? recommendations = shionSingingStream.recommendations;