+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - **Breaking**: Changed `HolodexClient::getVTuberClips` parameters to an optional `ChannelVideoFilter`
 - **Breaking**: Changed `HolodexClient::getChannelVideos` parameters to an optional `ChannelVideoFilter`
 - **Breaking**: Changed `HolodexClient::getVTuberCollabs` parameters to an optional `ChannelVideoFilter`
+- **Breaking**: Renamed `HolodexClient::getVideoMetadata` parameter `timestampComments` to `includeTimestampComments`
+- **Breaking**: Renamed `HolodexClient::getVideoMetadata` parameter `recommendationLanguages` to `languages`
 - **Breaking**: Removed `VideoWithCommentsList`
 - **Breaking**: Removed `VideoWithComments`, use `VideoFull` instead
 - **Breaking**: Removed `VideoMetadata`, use `VideoFull` instead